Remove spurious #to_s call.
Since `gets` returns a string (it must, otherwise we could not then call
chomp on it) and chomp also returns a string, we do not need to the call
to #to_s. While keeping it is harmless, one of the goals when
programming should be removing all but the necessary.