Move defined method into outer scope.
It's highly unidiomatic (meaning that no one ever does it) to have
nested method definitions. Mostly because the implicit reciever of the
inner method is the same as the outer, so both will be defined on the
same class or object, but the inner method will not be defined until
after the first call to the outer method.