Refactor to be less procedural and more Ruby-like.
While what you had worked, it was also very procedural, meaning that
anyone who wants to look at *what* your program does, must conern
themselves with how it does it. You described an algorithm for walking
over a string and changing any uppercase characters to be preceded with
an underscore, then downcased the entire string. And while it worked, I
was thinking about how you rename a file, not the fact that you. In a
perfect world, no one, not even you, cares *how* something is done, only
that it is done. I used a combination of regexp captures and the gsub
method's lesser-known block form to perform the same action that you did
without the procedural element "leaking" into the shown implementation.
I was able to do this by being familiar with the Ruby standard library,
but I still had to look up the block form on
http://ruby-doc.org/core-1.9.3/String.html#method-i-gsub

The big advantage is in clarity.